+++ ZcashLightClientKit/Transaction/WalletTransactionEncoder.swift
@@ -8,6 +8,7 @@
 import Foundation
 
 class WalletTransactionEncoder: TransactionEncoder {
+    
     var rustBackend: ZcashRustBackend.Type
     var repository: TransactionRepository
     var initializer: Initializer
@@ -17,7 +18,7 @@ class WalletTransactionEncoder: TransactionEncoder {
         self.initializer = initializer
     }
     
-    func createTransaction(spendingKey: String, zatoshi: Int64, to: String, memo: String?, from accountIndex: Int) throws -> EncodedTransaction {
+    func createTransaction(spendingKey: String, zatoshi: Int, to: String, memo: String?, from accountIndex: Int) throws -> EncodedTransaction {
         
         let txId = try createSpend(spendingKey: spendingKey, zatoshi: zatoshi, to: to, memo: memo, from: accountIndex)
         
@@ -35,7 +36,7 @@ class WalletTransactionEncoder: TransactionEncoder {
         }
     }
     
-    func createTransaction(spendingKey: String, zatoshi: Int64, to: String, memo: String?, from accountIndex: Int, result: @escaping TransactionEncoderResultBlock) {
+    func createTransaction(spendingKey: String, zatoshi: Int, to: String, memo: String?, from accountIndex: Int, result: @escaping TransactionEncoderResultBlock) {
         
         DispatchQueue.global(qos: .userInitiated).async { [weak self] in
             guard let self = self else { return }
@@ -47,7 +48,7 @@ class WalletTransactionEncoder: TransactionEncoder {
         } 
     }
     
-    func createSpend(spendingKey: String, zatoshi: Int64, to address: String, memo: String?, from accountIndex: Int) throws -> Int64 {
+    func createSpend(spendingKey: String, zatoshi: Int, to address: String, memo: String?, from accountIndex: Int) throws -> Int {
         guard ensureParams(spend: initializer.spendParamsURL, output: initializer.spendParamsURL),
             let spend = URL(string: initializer.spendParamsURL.path), let output = URL(string: initializer.outputParamsURL.path) else {
             throw TransactionEncoderError.missingParams
@@ -60,7 +61,7 @@ class WalletTransactionEncoder: TransactionEncoder {
             throw rustBackend.lastError() ?? RustWeldingError.genericError(message: "create spend failed")
         }
         
-        return txId
+        return Int(txId)
     }
     
     func ensureParams(spend: URL, output: URL) -> Bool {
